ボトムナビゲーションバーをリファクタリング
